Add an example of using nalgebra with the gauss-newton solver
The example uses dynamic matrices instead of static ones because `ArgminOp` only allows specifying the type of Jacobian/Hessian but not their transposes. Although it's possible to add 2 extra associated types for the transposes, it would be a breaking change; it's better to wait for when associated type defaults are stable.

Also, to get this working, the std::default::Default bound had to be removed from the solver's parameters because dynamic matrices in nalgebra don't implement it.