Fix clippy lint errors
There were a few errors to take care of here:

1. In `sr1_trustregion.rs`, there was an unnecessary borrow.

2. In many cases `assert_eq!(..., true)` or `assert_eq!(..., false)`
were being used. This is easily replaced by doing `assert!(...)`.

3. There were quite a number of checks on floating point numbers in
`iterstate.rs` that related to determining if a number was finite or
not. These are easily replaced by checking `value.is_infinite()`, along
with changing the `assert_eq!` to `assert!`.

4. There were a number of checks in the tests for the cost function that
compared two floating point values directly. This is always wrong. For
byte equality, it should compare `state.get_cost().to_ne_bytes()` to
`cost.to_ne_bytes()`. However, here the comparison is to validate that
these are in fact the same number. As the crate already depends on
`approx`, it is better to use `assert_relative_eq!` and set an epsilon.
Here I chose `f64::EPSILON`, which I would normally argue is a bad
choice, because it does not respect significance. However, I'm just
fixing lints, so I leave better choices of epsilon values to the
official maintainers.