Put serde dependency behind feature gate
This commit introduces the `serde1` feature gate which is part of the
default features. It removes the `serde` dependency, thus improving
compilation times. Removing `serde` removes the functionality to create
checkpoints, write parameter vectors to disk and logging to disk.
Also serde features of dependencies are linked to `serde1`.
`serde` may still be pulled in by other features such as `ndarrayl` and
`nalgebral`.

Since argmin uses `Serialize` and `DeserializeOwned` trait bounds on
many of its types, those were replace by `SerializeAlias` and
`DeserializeOwnedAlias` which changes its meaning depending on whether
`serde1` is on or off.

Other changes in this commit:

* Adapted CI accordingly (testing various features)
* Cleanup in the dependencies
  - removed `num`
  - added `num-complex` and `num-traits` instead
* added relevant features to examples in `Cargo.toml`
* `Serialize` and `Deserialize` are only derived when `serde1` is on
* Some tests are now also behind their appropriate feature gates (in
  particular `ndarrayl`)
* doctests where adapted so that they compile even when feature gates
  are off.
* docs were updated