Refactor math module into separate crate
This refactors the codebase into a workspace consisting of argmin and
argmin-math. argmin-math handles the math abstractions over other
computing crates such as `ndarray`, `nalgebra` and basic `Vec`s.
It includes the related math traits (otherwise the orphan rule would
prohibit this design). This means it needs to be a dependency of argmin.
It includes various features which allow one to select the computing
crates. It is also possible to turn all implementations off and only
include the traits.
The computing crates are support in various versions now, which should
make it easy to extend for future releases of these crates:

* ndarray
  - v0.15
  - v0.14
  - v0.13
* nalgebra
  - v0.30
  - v0.29

Serde support can also be toggled on and off.
The CI was adapted to the new structure and enhanced to cover various
combinations of features.

Other problems addressed in this commit:

* Fixed new clippy lint return_self_not_must_use
* Fixed new clippy lint nonminimal_bool
* Get current crate version in documentation from environment variables