More efficient ways of retrieving data from `IterState`
This adds multiple methods to `IterState` which allow one to either
move data, or copy data or get a reference to data when retrieving
parameter vectors, gradients, Jacobians, and Hessians.

In the past, only copying was possible. All solvers have been adapted to
use the new functions.

This also required `state` to be passed as a mutable reference into
`init` and `next_iter` of the solvers.