Removed visualizer feature and turned it into an example
The visualizer feature was a proof of concept and only works for 2D
problems. It wasn't updated in a very long time and does not seem to be
used a lot. Therefore it was decided to remove it and instead provide it
as an example for how to implement an `Observe`r.