Cleanup linesearch module and LineSearch trait
Cleanup of the linesearch module and documentation. Improved the
LineSearch trait by renaming the methods to something more intuitive.
Improved documentation of the LineSearch trait and added an example.
The individual line search modules are now private and the line searches
themselves are re-exported at the linesearch module level.
Acceptance conditions module remains public, but the conditions
themselves are not re-exported at the parent level anymore. The code
of the methods which use line searches was adapted accordingly.