Added float! macro for more convenient creation of float constants
When using generics with floats, creating a constant value from a
literal (such as `3.14`) is cumbersome:

`F::from_f64(3.14).unwrap()`

Not only does this take up quite a lot of space and is difficult to
read, it also comes with an `.unwrap()` which is not very pleasent on
the eyes (even though it is optimized away by the compiler).

The `float!` trait takes either a value:

`float!(<val>)`

which turns into this:

`F::from_f64(<val>).unwrap()`.

(This assumes that the generic parameter is `F` (which is typically the
case in argmin).)
Or it additionally takes a generic parameter as input:

`float!(<generic>, <val>)`

which turns into this:

`<generic>::from_f64(<val>).unwrap().

Overall this macro reduces the clutter in the code.