Cleanup of NewtonCG, better documentation and doc and unit tests
Method `curvature_threshold` renamed to `with_curvature_threshold`.

Hessian in subproblem is now stored as reference which avoids cloning of
the Hessian.

A missing initial parameter vector results in an Error.

Gradient and Hessian are now taken from state if available. This allows
users to provide Gradients and Hessians. If not availably they will be
computed.