Cleanup of PSO, documentation, doc and unit tests; Added PopulationState
The weights have been renamed to inertia weight and cognitive and social
acceleration factor. They now come with default values, which can be
changed via `with_inertia_factor`, `with_cognitive_factor` and
`with_social_factor`.

Initial creation of particles was redesigned to better allow for
potential future parallel computation.

Position<F> trait was removed, because it makes the code only slightly
easier to read but has the downside that it is exposed in the public
interface and therefore needs to be public as well.

Users can now initialize the solver with a population. The size of the
population must match `num_particles`. This is done by configuring the
state via the `population` method when setting up the executor.

Many `unwrap`s are now transformed to errors.

Many unit and doc tests were added.

`IterState` was replaced with the newly created `PopulationState` which
is tailored to population based solvers.
Population handling was removed from `IterState`.

The `particleswarm` and `observer` examples were adapted accordingly.