Parallel/bulk evaluation of problem functions
This commit implements `bulk_*` variants for `CostFunction`, `Operator`,
`Gradient`, `Jacobian` and `Hessian`. These methods process an array of
parameter vectors at once. If the new `rayon` feature is enabled, they
will be processed in parallel, if not, then they will be processed
sequentially. Even if `rayon` is enabled, parallel processing can
manually be turned off by implementing `parallelize` to return `false`.
Note though that bounds such as `Sync` and `Send` on parameter vectors,
gradients, jacobians, hessians, ... remain due to the `rayon`
dependency.

The `bulk_*` methods can be overwritten by the user and replaced with
other parallelization methods.

The methods are implemented by a macro called `bulk`.
They are also represented in `Problem`, because the number of function
evaluation needs to be counted correctly (one call to a `bulk_` method
of course leads to multiple cost function evaluations).

Since this needs to be thread safe, a couple of `Send` and `Sync` trait
bounds were needed. In order to not limit usability for builds without
the `rayon` feature, the `Send` and `Sync` traits are hidden behind
`SendAlias` and `SyncAlias`, respectively. With `rayon` enabled, they
will represent `Send` and `Sync`, if not, they are "empty". This is the
same approach as taken for `Serialize` and `DeserializeOwned`.

The only method so far which uses this bulk processing is Particle
Swarm Optimization, which was slightly adapted for parallel computation
of the cost function.

Furthermore the `trait_bounds` macro was removed as it was not used
anymore.