Cleanup of BFGS, added doc and unit tests, changed interface
The constructor of the BFGS method now does not take an initial inverse
Hessian anymore. Instead this needs to be provided via `configure` of
`Executor`, to be in line with how initial parameter vectors are
handled.

Documentation was improved and tolerance parameters are now explained.

The methods `with_tolerance_grad` and `with_tolerance_cost` now check
that the provided parameters are >=0.

Users can now provide the initial gradient and cost. If they are not
provided, they will be calculated.

Doc and unit tests were added.