Cleanup of DFP, added doc and unit tests, changed interface
The constructor of the DFP method now does not take an initial inverse
Hessian anymore. Instead this needs to be provided via `configure` of
`Executor`, to be in line with how initial parameter vectors are
handled.

Documentation was improved and the tolerance parameter is now explained.

The method `with_tolerance_grad` now checks that the paramter is >= 0.

Users can now provide the initial gradient and cost. If they are not
provided, they will be calculated.

Doc and unit tests were added.