Cleanup of SR1, added doc and unit tests, better documentation
Documentation was improved and tolerance parameters are now explained.

Method `r` was renamed to `with_denominator_factor`.

Values provided to the `with_*` methods are now checked to be within the
allowed bounds.

Initial Hessian was removed from the SR1 struct and is instead taken
from the state.

Users can now provide initial gradient and cost. Those will be
calculated if not provided.

Doc and unit tests were added.