Change links to solvers
Up until now, the README was created from the documentation in `lib.rs`.
Therefore it was necessary to link to the html page rather than using
the internal links in order for the links to work from the README. In
the future, the README will be an independent file and as such this is
not needed anymore. An advantage of this is that `cargo doc` will make
sure that the links are correct.