Fixed minmax for primitives in argmin-math
The minmax code wasn't part of the primitives module and therefore never
was compiled. Obviously the tests didn't run either. Once included in
the module tree, compilation failed because floats do not implement
`Ord`. Changed to a straight-forward implementation instead. Also removed
implementation for complex types because the other matrix backends don't
support those either and it's not entirely clear how to implement that
properly. Thanks to @Shreyan11 for reporting.